+++ baranov.pavel/P3/deleteDigits.hpp
@@ -1,0 +1,8 @@
+#ifndef DELETE_DIGITS_HPP
+#define DELETE_DIGITS_HPP
+#include <cstddef>
+
+void deleteDigits(const char * string, const size_t size, char * result);
+
+#endif
+
